\name{formato_robot}
\alias{formato_robot}
\title{Formato Robot}
\usage{
formato_robot()
}
\description{
Le da formato al fichero csv descargado de Prorealtime.
}
\arguments{
\item{robot}{dataset del csv descargado de Prorealtime}
\item{StartDay}{Desde que día quieres que coja datos}
\item{EndDay}{Hasta que día quieres que coja datos}
}

\examples{

}
